Split the vector plan verifier by responsibility
R5.2 of the transform cleanup plan: vector/verify/mod.rs becomes
model.rs (schema-v4 DTO vocabulary), error.rs (VectorPlanError +
impls), check.rs (independent checker + domain axioms),
fused_groups.rs, and checker_reachability.rs (documented as
producer-independent), behind an unchanged pub-use facade. Move only:
bodies byte-identical apart from pub(super) on sibling-crossed helpers
and per-file imports; all files under the 1,200-line threshold.

transform 385, vector_mode 35/35, clippy clean.
