+++ crates/transform/src/signal_fir/vector/verify/mod.rs
@@ -33,1951 +33,18 @@
 //! - **JSON (de)serialization / `plan_hash`** (R2 canonical-boundary work): a
 //!   plan is identified by its Rust type, not a runtime tag or hash.
 
-use ahash::{AHashMap, AHashSet};
-use std::collections::{BTreeMap, BTreeSet};
-use std::fmt;
+pub mod check;
+pub mod checker_reachability;
+pub mod error;
+pub mod fused_groups;
+pub mod model;
 
 pub use super::analysis::EffectAtom;
-use super::analysis::{DepKind, ForeignPurity, StateResource, effects_conflict};
-use super::decoration_verify::VerifiedDecorationCertificate;
-
-/// Current in-memory vector-plan schema. Version 4 records, per signal, the
-/// effects the signal performs itself alongside its transitive closure, so the
-/// event model can attribute an operation to its performer rather than to every
-/// carrier. Version 3 generalized fused serial groups from one recursive
-/// carrier to a canonical delayed-state carrier set.
-pub const VECTOR_PLAN_SCHEMA_VERSION: u32 = 4;
-
-/// `$defs/signalType`: the v1 value-type vocabulary (matches the Lean
-/// `ValueTy`). FIR widths / `FaustFloat` live in the routed-FIR layer, not
-/// here.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum ValueType {
-    Int,
-    Real,
-    /// Soundfile handle payload, derived from the signal shape exactly as the
-    /// prepared boundary derives `SimpleSigType::Sound`.
-    Sound,
-    Tuple(Vec<ValueType>),
-}
-
-/// `signalRecord.rate`.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum Rate {
-    Konst,
-    Block,
-    Samp,
-}
-
-/// `signalRecord.vectorability`.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum Vectorability {
-    Vect,
-    Scal,
-    TrueScal,
-}
-
-/// `$defs/placement`.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum Placement {
-    Control,
-    Inline,
-    Owned(u64),
-}
-
-/// `$defs/loopKind`.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum LoopKind {
-    Vectorizable,
-    Recursive(u64),
-    Island(u64),
-    /// One logical lane of a checked lockstep bundle. The lane loop identities
-    /// remain in the plan so the verifier can re-check independence; scheduling
-    /// and FIR assembly treat the bundle as one physical execution unit.
-    Lockstep {
-        width: u64,
-    },
-}
-
-/// `vecSafeWitness.witness_kind`.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum WitnessKind {
-    Pointwise,
-    SerialStateInternal,
-    ProvenIntrinsic,
-}
-
-/// `$defs/signalRecord`.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct SignalRecord {
-    pub signal_id: u64,
-    pub value_type: ValueType,
-    /// True only for checked symbolic recursion carriers with no runtime FIR value.
-    pub structural: bool,
-    pub rate: Rate,
-    pub vectorability: Vectorability,
-    pub clock_id: u64,
-    pub effects: Vec<EffectAtom>,
-    /// Effects this signal performs itself, a sorted subset of `effects`.
-    ///
-    /// The event model attributes an effect operation to the signal that
-    /// performs it. Attributing it to every transitive carrier instead invents
-    /// conflicts between signals that only contain the performer in their
-    /// subtree. `duplicable` keeps reading `effects`: a signal whose subtree
-    /// has any effect must not be duplicated, which is what keeps performers
-    /// unique.
-    pub direct_effects: Vec<EffectAtom>,
-    pub placement: Placement,
-    pub duplicable: bool,
-}
-
-/// `$defs/loopRecord`.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct LoopRecord {
-    pub loop_id: u64,
-    pub stable_name: String,
-    pub kind: LoopKind,
-    pub roots: Vec<u64>,
-    pub epoch_id: u64,
-}
-
-/// `$defs/epochRecord`.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct EpochRecord {
-    pub epoch_id: u64,
-    pub rank: u64,
-    pub loops: Vec<u64>,
-}
-
-/// `$defs/transportRecord`.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct TransportRecord {
-    pub transport_id: u64,
-    pub stable_name: String,
-    pub signal_id: u64,
-    pub producer_loop: u64,
-    pub consumer_loop: u64,
-    pub element_type: ValueType,
-    pub length: u64,
-    pub layout: TransportLayout,
-}
-
-/// Chunk-local transport layout. The external `compute` ABI remains planar;
-/// `Interleaved` is legal only for a checked lockstep bundle of matching width.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum TransportLayout {
-    Planar,
-    Interleaved(u64),
-}
-
-/// One explicit leaf correspondence in an isomorphism witness.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord)]
-pub struct IsoLeafMapping {
-    pub representative_signal_id: u64,
-    pub lane_signal_id: u64,
-}
-
-/// Root-level isomorphism evidence for one non-representative lane.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct IsoRootWitness {
-    pub representative_root: u64,
-    pub lane_root: u64,
-    pub shape_hash: u64,
-    pub leaf_mapping: Vec<IsoLeafMapping>,
-}
-
-/// One logical lane retained inside a lockstep bundle certificate.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct LockstepLaneRecord {
-    pub loop_id: u64,
-    pub recursion_group: u64,
-    pub roots: Vec<IsoRootWitness>,
-}
-
-/// Canonical lockstep bundle. `member_loop_ids` are scheduled as one unit and
-/// assembled under one physical sample loop, while remaining explicit in the
-/// plan for legality checks.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct LockstepBundleRecord {
-    pub bundle_id: u64,
-    pub representative_loop_id: u64,
-    pub member_loop_ids: Vec<u64>,
-    pub lanes: Vec<LockstepLaneRecord>,
-}
-
-/// `$defs/fusedSerialGroupRecord`.
-///
-/// A group preserves the original loop identities while requiring their
-/// delayed state work to be emitted as one serial per-sample unit. Generic
-/// delayed-state and recursive owners may share one group; `owner_loop_id` is
-/// its canonical minimum carrier owner. `state_carrier_signal_ids` is the
-/// non-empty, strictly ascending, complete set of positive-delay carriers,
-/// while `state_write_signal_ids` covers every carrier and recursive member.
-/// An `internal_transport_id` retains a planned transport identity whose
-/// access is rematerialized as a scalar value inside that unit. All members
-/// and grouped signals must have one exact clock id.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct FusedSerialGroupRecord {
-    pub group_id: u64,
-    pub owner_loop_id: u64,
-    pub member_loop_ids: Vec<u64>,
-    pub state_carrier_signal_ids: Vec<u64>,
-    pub delayed_read_signal_ids: Vec<u64>,
-    pub state_write_signal_ids: Vec<u64>,
-    pub internal_transport_ids: Vec<u64>,
-    pub output_or_transport_roots: Vec<u64>,
-}
-
-/// `$defs/vecSafeWitness`.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct VecSafeWitness {
-    pub loop_id: u64,
-    pub witness_kind: WitnessKind,
-}
-
-/// A `data` or `effect` edge over loop ids (`consumer -> dependency`:
-/// dependency runs first).
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord)]
-pub struct LoopEdge {
-    pub consumer: u64,
-    pub dependency: u64,
-}
-
-/// `$defs/vectorPlan`.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct VectorPlan {
-    pub schema_version: u32,
-    pub vec_size: u64,
-    pub signals: Vec<SignalRecord>,
-    pub loops: Vec<LoopRecord>,
-    pub epochs: Vec<EpochRecord>,
-    pub transports: Vec<TransportRecord>,
-    pub data_edges: Vec<LoopEdge>,
-    pub effect_edges: Vec<LoopEdge>,
-    pub vec_safe_witnesses: Vec<VecSafeWitness>,
-    pub fused_serial_groups: Vec<FusedSerialGroupRecord>,
-    pub lockstep_bundles: Vec<LockstepBundleRecord>,
-}
-
-/// Why [`verify_vector_plan`] rejected a plan. One variant per checked
-/// obligation, so each has a demonstrated rejecting mutation (plan §8).
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum VectorPlanError {
-    /// The verifier accepts only the exact v2 schema.
-    UnsupportedSchema { found: u32 },
-    /// `vec_size` must be positive.
-    VecSizeZero,
-    /// A set-like array is not in its required canonical order (also catches
-    /// duplicates, since canonical order is *strictly* ascending).
-    NotCanonical { what: &'static str, at: usize },
-    /// A loop appears in more than one epoch, or a plan loop is in none.
-    EpochCoverageMismatch { loop_id: u64 },
-    /// An epoch lists a loop id that is not a plan loop.
-    EpochLoopUnknown { epoch_id: u64, loop_id: u64 },
-    /// A signal placed `Owned(l)` is absent from `l`'s roots.
-    OwnedSignalNotRoot { signal_id: u64, loop_id: u64 },
-    /// A root of loop `l` is not placed `Owned(l)`.
-    RootWithoutOwnership { signal_id: u64, loop_id: u64 },
-    /// A root references a signal id that is not a plan signal.
-    RootUnknownSignal { loop_id: u64, signal_id: u64 },
-    /// An `Inline`-placed signal is not `duplicable`.
-    InlineNotDuplicable { signal_id: u64 },
-    /// The producer-supplied `duplicable` bit disagrees with the effect facts.
-    DuplicabilityMismatch { signal_id: u64 },
-    /// A loop's redundant `epoch_id` disagrees with canonical epoch membership.
-    LoopEpochMismatch {
-        loop_id: u64,
-        declared: u64,
-        actual: u64,
-    },
-    /// A data/effect edge references a loop id that is not a plan loop.
-    EdgeEndpointUnknown { edge: LoopEdge, missing: u64 },
-    /// A loop depends on itself (an instantaneous self-edge).
-    LoopSelfEdge { loop_id: u64 },
-    /// The induced graph of one epoch contains a cycle.
-    EpochNotAcyclic { epoch_id: u64, remaining: Vec<u64> },
-    /// Two loops have conflicting effects but neither is ordered before the
-    /// other by the combined data/effect relation.
-    UnorderedEffectConflict { left: u64, right: u64 },
-    /// A transport's producer and consumer loops are the same.
-    TransportSelfLoop { transport_id: u64 },
-    /// A transport's element type does not equal its signal's value type.
-    TransportTypeMismatch { transport_id: u64 },
-    /// A transport's array length does not equal `vec_size`.
-    TransportLengthMismatch { transport_id: u64 },
-    /// An interleaved transport has no matching lockstep width.
-    TransportLayoutMismatch { transport_id: u64 },
-    /// A lockstep bundle has fewer than two lanes or inconsistent width.
-    LockstepWidthMismatch { bundle_id: u64 },
-    /// A lockstep bundle references a missing loop or repeats a member.
-    LockstepMemberMismatch { bundle_id: u64, loop_id: u64 },
-    /// A lane record does not correspond exactly to one bundle member.
-    LockstepLaneMismatch { bundle_id: u64, loop_id: u64 },
-    /// Two candidate lanes are connected in the epoch dependence graph.
-    LockstepDependentLanes {
-        bundle_id: u64,
-        left: u64,
-        right: u64,
-    },
-    /// Two candidate lanes do not share the same epoch and clock.
-    LockstepDomainMismatch {
-        bundle_id: u64,
-        left: u64,
-        right: u64,
-    },
-    /// Two candidate lanes have non-commuting effects.
-    LockstepEffectConflict {
-        bundle_id: u64,
-        left: u64,
-        right: u64,
-    },
-    /// A root/leaf witness is not canonical or references signals outside its
-    /// declared lane roots. Prepared-tree shape is checked by the second gate.
-    LockstepIsoWitnessMismatch { bundle_id: u64, loop_id: u64 },
-    /// A cross-epoch edge whose dependency epoch has a strictly greater rank
-    /// than its consumer epoch (a barrier run backwards).
-    BarrierViolation { edge: LoopEdge },
-    /// A `Recursive`/`Island` loop carries a `pointwise` witness, or is
-    /// otherwise asserted vector-safe in a way that contradicts its serial
-    /// kind.
-    SerialLoopNotSerial { loop_id: u64 },
-    /// A `Vectorizable` loop has no `VecSafe` witness.
-    VectorizableWithoutWitness { loop_id: u64 },
-    /// A vectorizable loop's roots do not satisfy the concrete `VecSafe` rule.
-    VectorizableNotSafe { loop_id: u64 },
-    /// A `VecSafe` witness references a loop id that is not a plan loop.
-    WitnessUnknownLoop { loop_id: u64 },
-    /// A transport references a signal or loop id that is not in the plan.
-    TransportUnknownRef { transport_id: u64, missing: u64 },
-    /// A required set-like fused-group field is empty.
-    FusedGroupEmpty { group_id: u64, what: &'static str },
-    /// A fused group references a loop absent from the plan.
-    FusedGroupUnknownLoop { group_id: u64, loop_id: u64 },
-    /// The owner is not included in the group's members.
-    FusedGroupOwnerNotMember { group_id: u64, owner_loop_id: u64 },
-    /// One loop belongs to two fused serial groups.
-    FusedGroupLoopOverlap { loop_id: u64 },
-    /// A fused group references a signal absent from the plan.
-    FusedGroupUnknownSignal { group_id: u64, signal_id: u64 },
-    /// A grouped signal is not owned by one of the group's member loops.
-    FusedGroupSignalOutside { group_id: u64, signal_id: u64 },
-    /// A rematerialized transport id is absent from the plan.
-    FusedGroupUnknownTransport { group_id: u64, transport_id: u64 },
-    /// A rematerialized transport does not stay within its group.
-    FusedGroupTransportOutside { group_id: u64, transport_id: u64 },
-    /// A rematerialized transport does not carry one of the certified delayed
-    /// reads.
-    FusedGroupTransportNotDelayedRead { group_id: u64, transport_id: u64 },
-    /// The selected owner is not the canonical owner of a certified carrier.
-    FusedGroupOwnerNotStateCarrier { group_id: u64, owner_loop_id: u64 },
-    /// The decoration facts do not identify the declared carrier as delayed
-    /// state.
-    FusedGroupCarrierNotDelayedState { group_id: u64, signal_id: u64 },
-    /// A delayed read lacks the declared `DepKind::Delayed` carrier edge.
-    FusedGroupDelayedDependencyMissing { group_id: u64, signal_id: u64 },
-    /// A same-sample path from a delayed read to its recursive writer leaves
-    /// the fused serial group.
-    FusedGroupPathIncomplete { group_id: u64, loop_id: u64 },
-    /// A declared state writer does not match its recursive member loop.
-    FusedGroupStateWriterMismatch { group_id: u64, signal_id: u64 },
-    /// A recursive member loop has no matching projection writer in the group.
-    FusedGroupRecursiveMemberMissingWriter { group_id: u64, loop_id: u64 },
-    /// Grouped signals or loops cross incompatible clock domains.
-    FusedGroupClockMismatch { group_id: u64 },
-    /// An active chunk transport still materializes a delayed read internally.
-    FusedGroupDangerousTransportPresent { group_id: u64, transport_id: u64 },
-    /// An independently reconstructed immediate-delay crossing is uncovered.
-    FusedGroupDangerousCrossingMissing { producer: u64, consumer: u64 },
-}
-
-impl fmt::Display for VectorPlanError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::UnsupportedSchema { found } => {
-                write!(f, "unsupported vector-plan schema {found}")
-            }
-            Self::VecSizeZero => write!(f, "vec_size must be positive"),
-            Self::NotCanonical { what, at } => {
-                write!(f, "{what} is not strictly ascending at index {at}")
-            }
-            Self::EpochCoverageMismatch { loop_id } => {
-                write!(f, "loop {loop_id} is not owned by exactly one epoch")
-            }
-            Self::EpochLoopUnknown { epoch_id, loop_id } => {
-                write!(f, "epoch {epoch_id} lists unknown loop {loop_id}")
-            }
-            Self::OwnedSignalNotRoot { signal_id, loop_id } => write!(
-                f,
-                "signal {signal_id} is Owned({loop_id}) but not a root of that loop"
-            ),
-            Self::RootWithoutOwnership { signal_id, loop_id } => write!(
-                f,
-                "signal {signal_id} is a root of loop {loop_id} but not placed Owned({loop_id})"
-            ),
-            Self::RootUnknownSignal { loop_id, signal_id } => {
-                write!(f, "loop {loop_id} root {signal_id} is not a plan signal")
-            }
-            Self::InlineNotDuplicable { signal_id } => {
-                write!(f, "Inline signal {signal_id} is not duplicable")
-            }
-            Self::DuplicabilityMismatch { signal_id } => write!(
-                f,
-                "signal {signal_id} duplicable bit disagrees with its effects"
-            ),
-            Self::LoopEpochMismatch {
-                loop_id,
-                declared,
-                actual,
-            } => write!(
-                f,
-                "loop {loop_id} declares epoch {declared} but belongs to epoch {actual}"
-            ),
-            Self::EdgeEndpointUnknown { edge, missing } => {
-                write!(f, "edge {edge:?} references unknown loop {missing}")
-            }
-            Self::LoopSelfEdge { loop_id } => write!(f, "loop {loop_id} depends on itself"),
-            Self::EpochNotAcyclic {
-                epoch_id,
-                remaining,
-            } => write!(
-                f,
-                "epoch {epoch_id} induced graph has a cycle: {remaining:?}"
-            ),
-            Self::UnorderedEffectConflict { left, right } => write!(
-                f,
-                "loops {left} and {right} have conflicting unordered effects"
-            ),
-            Self::TransportSelfLoop { transport_id } => {
-                write!(f, "transport {transport_id} producer == consumer")
-            }
-            Self::TransportTypeMismatch { transport_id } => write!(
-                f,
-                "transport {transport_id} element type != its signal's value type"
-            ),
-            Self::TransportLengthMismatch { transport_id } => {
-                write!(f, "transport {transport_id} length != vec_size")
-            }
-            Self::TransportLayoutMismatch { transport_id } => {
-                write!(f, "transport {transport_id} has an invalid lockstep layout")
-            }
-            Self::LockstepWidthMismatch { bundle_id } => {
-                write!(f, "lockstep bundle {bundle_id} has an inconsistent width")
-            }
-            Self::LockstepMemberMismatch { bundle_id, loop_id } => write!(
-                f,
-                "lockstep bundle {bundle_id} has invalid member loop {loop_id}"
-            ),
-            Self::LockstepLaneMismatch { bundle_id, loop_id } => write!(
-                f,
-                "lockstep bundle {bundle_id} has invalid lane loop {loop_id}"
-            ),
-            Self::LockstepDependentLanes {
-                bundle_id,
-                left,
-                right,
-            } => write!(
-                f,
-                "lockstep bundle {bundle_id} lanes {left} and {right} are dependency-connected"
-            ),
-            Self::LockstepDomainMismatch {
-                bundle_id,
-                left,
-                right,
-            } => write!(
-                f,
-                "lockstep bundle {bundle_id} lanes {left} and {right} disagree on epoch or clock"
-            ),
-            Self::LockstepEffectConflict {
-                bundle_id,
-                left,
-                right,
-            } => write!(
-                f,
-                "lockstep bundle {bundle_id} lanes {left} and {right} have conflicting effects"
-            ),
-            Self::LockstepIsoWitnessMismatch { bundle_id, loop_id } => write!(
-                f,
-                "lockstep bundle {bundle_id} lane {loop_id} has an invalid isomorphism witness"
-            ),
-            Self::BarrierViolation { edge } => {
-                write!(f, "cross-epoch edge {edge:?} runs a barrier backwards")
-            }
-            Self::SerialLoopNotSerial { loop_id } => {
-                write!(
-                    f,
-                    "serial (recursive/island) loop {loop_id} asserted vector-safe"
-                )
-            }
-            Self::VectorizableWithoutWitness { loop_id } => {
-                write!(f, "vectorizable loop {loop_id} has no VecSafe witness")
-            }
-            Self::VectorizableNotSafe { loop_id } => {
-                write!(f, "vectorizable loop {loop_id} does not satisfy VecSafe")
-            }
-            Self::WitnessUnknownLoop { loop_id } => {
-                write!(f, "VecSafe witness references unknown loop {loop_id}")
-            }
-            Self::TransportUnknownRef {
-                transport_id,
-                missing,
-            } => write!(
-                f,
-                "transport {transport_id} references unknown id {missing}"
-            ),
-            Self::FusedGroupEmpty { group_id, what } => {
-                write!(f, "fused group {group_id} has empty {what}")
-            }
-            Self::FusedGroupUnknownLoop { group_id, loop_id } => {
-                write!(
-                    f,
-                    "fused group {group_id} references unknown loop {loop_id}"
-                )
-            }
-            Self::FusedGroupOwnerNotMember {
-                group_id,
-                owner_loop_id,
-            } => write!(
-                f,
-                "fused group {group_id} owner loop {owner_loop_id} is not a member"
-            ),
-            Self::FusedGroupLoopOverlap { loop_id } => {
-                write!(f, "loop {loop_id} belongs to more than one fused group")
-            }
-            Self::FusedGroupUnknownSignal {
-                group_id,
-                signal_id,
-            } => write!(
-                f,
-                "fused group {group_id} references unknown signal {signal_id}"
-            ),
-            Self::FusedGroupSignalOutside {
-                group_id,
-                signal_id,
-            } => write!(
-                f,
-                "fused group {group_id} signal {signal_id} is outside its member loops"
-            ),
-            Self::FusedGroupUnknownTransport {
-                group_id,
-                transport_id,
-            } => write!(
-                f,
-                "fused group {group_id} references unknown transport {transport_id}"
-            ),
-            Self::FusedGroupTransportOutside {
-                group_id,
-                transport_id,
-            } => write!(
-                f,
-                "fused group {group_id} transport {transport_id} crosses its member boundary"
-            ),
-            Self::FusedGroupTransportNotDelayedRead {
-                group_id,
-                transport_id,
-            } => write!(
-                f,
-                "fused group {group_id} transport {transport_id} is not a delayed read"
-            ),
-            Self::FusedGroupOwnerNotStateCarrier {
-                group_id,
-                owner_loop_id,
-            } => write!(
-                f,
-                "fused group {group_id} owner loop {owner_loop_id} is not the canonical state-carrier owner"
-            ),
-            Self::FusedGroupCarrierNotDelayedState {
-                group_id,
-                signal_id,
-            } => write!(
-                f,
-                "fused group {group_id} carrier {signal_id} is not certified delayed state"
-            ),
-            Self::FusedGroupDelayedDependencyMissing {
-                group_id,
-                signal_id,
-            } => write!(
-                f,
-                "fused group {group_id} read {signal_id} lacks its delayed carrier dependency"
-            ),
-            Self::FusedGroupPathIncomplete { group_id, loop_id } => {
-                write!(f, "fused group {group_id} omits data-path loop {loop_id}")
-            }
-            Self::FusedGroupStateWriterMismatch {
-                group_id,
-                signal_id,
-            } => write!(
-                f,
-                "fused group {group_id} writer {signal_id} does not match a recursive member loop"
-            ),
-            Self::FusedGroupRecursiveMemberMissingWriter { group_id, loop_id } => write!(
-                f,
-                "fused group {group_id} recursive member loop {loop_id} has no state writer"
-            ),
-            Self::FusedGroupClockMismatch { group_id } => {
-                write!(f, "fused group {group_id} crosses clock domains")
-            }
-            Self::FusedGroupDangerousTransportPresent {
-                group_id,
-                transport_id,
-            } => write!(
-                f,
-                "fused group {group_id} still materializes delayed read transport {transport_id}"
-            ),
-            Self::FusedGroupDangerousCrossingMissing { producer, consumer } => write!(
-                f,
-                "state-mediated immediate delay crossing {producer} -> {consumer} is not covered by a fused group"
-            ),
-        }
-    }
-}
-
-impl std::error::Error for VectorPlanError {}
-
-fn strictly_ascending<T: Ord>(items: &[T]) -> Result<(), usize> {
-    for i in 1..items.len() {
-        if items[i - 1] >= items[i] {
-            return Err(i);
-        }
-    }
-    Ok(())
-}
-
-/// Concrete Lean `duplicableEffectsB`: only an empty effect set or pure
-/// foreign calls can be recomputed in several loop regions.
-#[must_use]
-pub(crate) fn effects_duplicable(effects: &[EffectAtom]) -> bool {
-    effects.iter().all(|effect| {
-        matches!(
-            effect,
-            EffectAtom::Foreign {
-                purity: ForeignPurity::Pure,
-                ..
-            }
-        )
-    })
-}
-
-/// Concrete Lean `sampleReorderableB`: loop-carried state is the local
-/// per-sample vectorization blocker. Other effect conflicts are ordered or
-/// co-located by the plan's separate effect relation.
-#[must_use]
-pub(crate) fn effects_sample_reorderable(effects: &[EffectAtom]) -> bool {
-    !effects
-        .iter()
-        .any(|effect| matches!(effect, EffectAtom::ReadState(_) | EffectAtom::WriteState(_)))
-}
-
-/// Independent verifier for a [`VectorPlan`] (plan §5.5/§5.10
-/// `verify_vector_plan`; Lean `VectorPlanCertificate`). Re-derives every
-/// invariant from the plan's own fields; never runs a planner.
-///
-/// # Errors
-/// The first [`VectorPlanError`] found (checks ordered so identity/coverage
-/// problems surface before the graph/transport checks that assume them).
-pub fn verify_vector_plan(plan: &VectorPlan) -> Result<(), VectorPlanError> {
-    if plan.schema_version != VECTOR_PLAN_SCHEMA_VERSION {
-        return Err(VectorPlanError::UnsupportedSchema {
-            found: plan.schema_version,
-        });
-    }
-    if plan.vec_size == 0 {
-        return Err(VectorPlanError::VecSizeZero);
-    }
-
-    // ── Canonical set orders (also enforce uniqueness). ──────────────────
-    let signal_ids: Vec<u64> = plan.signals.iter().map(|s| s.signal_id).collect();
-    strictly_ascending(&signal_ids).map_err(|at| VectorPlanError::NotCanonical {
-        what: "signals",
-        at,
-    })?;
-    let loop_ids: Vec<u64> = plan.loops.iter().map(|l| l.loop_id).collect();
-    strictly_ascending(&loop_ids)
-        .map_err(|at| VectorPlanError::NotCanonical { what: "loops", at })?;
-    let epoch_keys: Vec<(u64, u64)> = plan.epochs.iter().map(|e| (e.rank, e.epoch_id)).collect();
-    strictly_ascending(&epoch_keys)
-        .map_err(|at| VectorPlanError::NotCanonical { what: "epochs", at })?;
-    let transport_ids: Vec<u64> = plan.transports.iter().map(|t| t.transport_id).collect();
-    strictly_ascending(&transport_ids).map_err(|at| VectorPlanError::NotCanonical {
-        what: "transports",
-        at,
-    })?;
-    strictly_ascending(&plan.data_edges).map_err(|at| VectorPlanError::NotCanonical {
-        what: "data_edges",
-        at,
-    })?;
-    strictly_ascending(&plan.effect_edges).map_err(|at| VectorPlanError::NotCanonical {
-        what: "effect_edges",
-        at,
-    })?;
-    let witness_ids: Vec<u64> = plan
-        .vec_safe_witnesses
-        .iter()
-        .map(|witness| witness.loop_id)
-        .collect();
-    strictly_ascending(&witness_ids).map_err(|at| VectorPlanError::NotCanonical {
-        what: "vec_safe_witnesses",
-        at,
-    })?;
-    let fused_group_ids = plan
-        .fused_serial_groups
-        .iter()
-        .map(|group| group.group_id)
-        .collect::<Vec<_>>();
-    strictly_ascending(&fused_group_ids).map_err(|at| VectorPlanError::NotCanonical {
-        what: "fused_serial_groups",
-        at,
-    })?;
-    let lockstep_bundle_ids = plan
-        .lockstep_bundles
-        .iter()
-        .map(|bundle| bundle.bundle_id)
-        .collect::<Vec<_>>();
-    strictly_ascending(&lockstep_bundle_ids).map_err(|at| VectorPlanError::NotCanonical {
-        what: "lockstep_bundles",
-        at,
-    })?;
-
-    let signal_set: AHashSet<u64> = signal_ids.iter().copied().collect();
-    let loop_set: AHashSet<u64> = loop_ids.iter().copied().collect();
-    let signal_by_id: AHashMap<u64, &SignalRecord> =
-        plan.signals.iter().map(|s| (s.signal_id, s)).collect();
-    let loop_by_id: AHashMap<u64, &LoopRecord> =
-        plan.loops.iter().map(|l| (l.loop_id, l)).collect();
-
-    // ── Lockstep finite shape. Prepared-signal skeletons are re-traversed by
-    // `verify_lockstep_isomorphism`; this plan-local gate checks every graph,
-    // effect, epoch, ownership, and canonical witness obligation first.
-    for loop_record in &plan.loops {
-        if let Some(&signal_id) = loop_record
-            .roots
-            .iter()
-            .find(|signal_id| !signal_set.contains(signal_id))
-        {
-            return Err(VectorPlanError::RootUnknownSignal {
-                loop_id: loop_record.loop_id,
-                signal_id,
-            });
-        }
-    }
-    for edge in plan.data_edges.iter().chain(&plan.effect_edges) {
-        if !loop_set.contains(&edge.consumer) {
-            return Err(VectorPlanError::EdgeEndpointUnknown {
-                edge: *edge,
-                missing: edge.consumer,
-            });
-        }
-        if !loop_set.contains(&edge.dependency) {
-            return Err(VectorPlanError::EdgeEndpointUnknown {
-                edge: *edge,
-                missing: edge.dependency,
-            });
-        }
-    }
-    let reachability = CheckedReachability::new(plan);
-    let effects_by_loop = plan
-        .loops
-        .iter()
-        .map(|loop_record| {
-            (
-                loop_record.loop_id,
-                CheckedEffectConflictSummary::new(&signal_by_id, loop_record),
-            )
-        })
-        .collect::<AHashMap<_, _>>();
-    let mut bundled_loops = AHashSet::new();
-    for bundle in &plan.lockstep_bundles {
-        if bundle.member_loop_ids.len() < 2 || bundle.lanes.len() != bundle.member_loop_ids.len() {
-            return Err(VectorPlanError::LockstepWidthMismatch {
-                bundle_id: bundle.bundle_id,
-            });
-        }
-        strictly_ascending(&bundle.member_loop_ids).map_err(|at| {
-            VectorPlanError::NotCanonical {
-                what: "lockstep.member_loop_ids",
-                at,
-            }
-        })?;
-        if bundle.member_loop_ids.first().copied() != Some(bundle.representative_loop_id) {
-            return Err(VectorPlanError::LockstepMemberMismatch {
-                bundle_id: bundle.bundle_id,
-                loop_id: bundle.representative_loop_id,
-            });
-        }
-        let lane_loop_ids = bundle
-            .lanes
-            .iter()
-            .map(|lane| lane.loop_id)
-            .collect::<Vec<_>>();
-        if lane_loop_ids != bundle.member_loop_ids {
-            return Err(VectorPlanError::LockstepLaneMismatch {
-                bundle_id: bundle.bundle_id,
-                loop_id: lane_loop_ids
-                    .iter()
-                    .zip(&bundle.member_loop_ids)
-                    .find_map(|(actual, expected)| (actual != expected).then_some(*actual))
-                    .unwrap_or(bundle.representative_loop_id),
-            });
-        }
-        let width =
-            u64::try_from(bundle.member_loop_ids.len()).expect("lockstep member count fits u64");
-        let representative = loop_by_id.get(&bundle.representative_loop_id).ok_or(
-            VectorPlanError::LockstepMemberMismatch {
-                bundle_id: bundle.bundle_id,
-                loop_id: bundle.representative_loop_id,
-            },
-        )?;
-        for lane in &bundle.lanes {
-            let Some(loop_record) = loop_by_id.get(&lane.loop_id).copied() else {
-                return Err(VectorPlanError::LockstepMemberMismatch {
-                    bundle_id: bundle.bundle_id,
-                    loop_id: lane.loop_id,
-                });
-            };
-            if loop_record.kind != (LoopKind::Lockstep { width }) {
-                return Err(VectorPlanError::LockstepWidthMismatch {
-                    bundle_id: bundle.bundle_id,
-                });
-            }
-            if !bundled_loops.insert(lane.loop_id)
-                || loop_record.epoch_id != representative.epoch_id
-                || lane.roots.len() != loop_record.roots.len()
-                || lane
-                    .roots
-                    .iter()
-                    .map(|root| root.lane_root)
-                    .ne(loop_record.roots.iter().copied())
-                || lane
-                    .roots
-                    .iter()
-                    .map(|root| root.representative_root)
-                    .ne(representative.roots.iter().copied())
-            {
-                return Err(VectorPlanError::LockstepLaneMismatch {
-                    bundle_id: bundle.bundle_id,
-                    loop_id: lane.loop_id,
-                });
-            }
-            for root in &lane.roots {
-                if root.shape_hash == 0
-                    || !signal_set.contains(&root.representative_root)
-                    || !signal_set.contains(&root.lane_root)
-                    || strictly_ascending(&root.leaf_mapping).is_err()
-                    || root.leaf_mapping.iter().any(|mapping| {
-                        !signal_set.contains(&mapping.representative_signal_id)
-                            || !signal_set.contains(&mapping.lane_signal_id)
-                    })
-                {
-                    return Err(VectorPlanError::LockstepIsoWitnessMismatch {
-                        bundle_id: bundle.bundle_id,
-                        loop_id: lane.loop_id,
-                    });
-                }
-            }
-        }
-        for (index, &left) in bundle.member_loop_ids.iter().enumerate() {
-            for &right in &bundle.member_loop_ids[index + 1..] {
-                if reachability.reaches(left, right) || reachability.reaches(right, left) {
-                    return Err(VectorPlanError::LockstepDependentLanes {
-                        bundle_id: bundle.bundle_id,
-                        left,
-                        right,
-                    });
-                }
-                let left_loop = loop_by_id[&left];
-                let right_loop = loop_by_id[&right];
-                let left_clocks = left_loop
-                    .roots
-                    .iter()
-                    .map(|root| signal_by_id[root].clock_id)
-                    .collect::<AHashSet<_>>();
-                let right_clocks = right_loop
-                    .roots
-                    .iter()
-                    .map(|root| signal_by_id[root].clock_id)
-                    .collect::<AHashSet<_>>();
-                if left_loop.epoch_id != right_loop.epoch_id || left_clocks != right_clocks {
-                    return Err(VectorPlanError::LockstepDomainMismatch {
-                        bundle_id: bundle.bundle_id,
-                        left,
-                        right,
-                    });
-                }
-                if effects_by_loop[&left].conflicts(&effects_by_loop[&right]) {
-                    return Err(VectorPlanError::LockstepEffectConflict {
-                        bundle_id: bundle.bundle_id,
-                        left,
-                        right,
-                    });
-                }
-            }
-        }
-    }
-    for loop_record in &plan.loops {
-        if matches!(loop_record.kind, LoopKind::Lockstep { .. })
-            != bundled_loops.contains(&loop_record.loop_id)
-        {
-            return Err(VectorPlanError::LockstepMemberMismatch {
-                bundle_id: plan
-                    .lockstep_bundles
-                    .iter()
-                    .find(|bundle| bundle.member_loop_ids.contains(&loop_record.loop_id))
-                    .map_or(u64::MAX, |bundle| bundle.bundle_id),
-                loop_id: loop_record.loop_id,
-            });
-        }
-    }
-
-    // ── Fused-group finite shape. Semantic delay/recursion facts are checked
-    // independently by `verify_fused_serial_groups` against decorations.
-    let transport_by_id = plan
-        .transports
-        .iter()
-        .map(|transport| (transport.transport_id, transport))
-        .collect::<AHashMap<_, _>>();
-    let mut fused_loop_owner = AHashSet::new();
-    for group in &plan.fused_serial_groups {
-        for (what, ids) in [
-            ("member_loop_ids", group.member_loop_ids.as_slice()),
-            (
-                "state_carrier_signal_ids",
-                group.state_carrier_signal_ids.as_slice(),
-            ),
-            (
-                "delayed_read_signal_ids",
-                group.delayed_read_signal_ids.as_slice(),
-            ),
-            (
-                "state_write_signal_ids",
-                group.state_write_signal_ids.as_slice(),
-            ),
-            (
-                "output_or_transport_roots",
-                group.output_or_transport_roots.as_slice(),
-            ),
-        ] {
-            if ids.is_empty() {
-                return Err(VectorPlanError::FusedGroupEmpty {
-                    group_id: group.group_id,
-                    what,
-                });
-            }
-            strictly_ascending(ids).map_err(|at| VectorPlanError::NotCanonical { what, at })?;
-        }
-        strictly_ascending(&group.internal_transport_ids).map_err(|at| {
-            VectorPlanError::NotCanonical {
-                what: "internal_transport_ids",
-                at,
-            }
-        })?;
-        for &loop_id in &group.member_loop_ids {
-            if !loop_set.contains(&loop_id) {
-                return Err(VectorPlanError::FusedGroupUnknownLoop {
-                    group_id: group.group_id,
-                    loop_id,
-                });
-            }
-            if !fused_loop_owner.insert(loop_id) {
-                return Err(VectorPlanError::FusedGroupLoopOverlap { loop_id });
-            }
-        }
-        if group
-            .member_loop_ids
-            .binary_search(&group.owner_loop_id)
-            .is_err()
-        {
-            return Err(VectorPlanError::FusedGroupOwnerNotMember {
-                group_id: group.group_id,
-                owner_loop_id: group.owner_loop_id,
-            });
-        }
-        for &signal_id in group
-            .state_carrier_signal_ids
-            .iter()
-            .chain(&group.delayed_read_signal_ids)
-            .chain(&group.state_write_signal_ids)
-            .chain(&group.output_or_transport_roots)
-        {
-            let Some(signal) = signal_by_id.get(&signal_id) else {
-                return Err(VectorPlanError::FusedGroupUnknownSignal {
-                    group_id: group.group_id,
-                    signal_id,
-                });
-            };
-            let Placement::Owned(owner) = signal.placement else {
-                return Err(VectorPlanError::FusedGroupSignalOutside {
-                    group_id: group.group_id,
-                    signal_id,
-                });
-            };
-            if group.member_loop_ids.binary_search(&owner).is_err() {
-                return Err(VectorPlanError::FusedGroupSignalOutside {
-                    group_id: group.group_id,
-                    signal_id,
-                });
-            }
-        }
-        if !group.state_carrier_signal_ids.iter().any(|signal_id| {
-            signal_by_id
-                .get(signal_id)
-                .is_some_and(|signal| signal.placement == Placement::Owned(group.owner_loop_id))
-        }) {
-            return Err(VectorPlanError::FusedGroupOwnerNotStateCarrier {
-                group_id: group.group_id,
-                owner_loop_id: group.owner_loop_id,
-            });
-        }
-        for &transport_id in &group.internal_transport_ids {
-            let Some(transport) = transport_by_id.get(&transport_id) else {
-                return Err(VectorPlanError::FusedGroupUnknownTransport {
-                    group_id: group.group_id,
-                    transport_id,
-                });
-            };
-            if group
-                .member_loop_ids
-                .binary_search(&transport.producer_loop)
-                .is_err()
-                || group
-                    .member_loop_ids
-                    .binary_search(&transport.consumer_loop)
-                    .is_err()
-            {
-                return Err(VectorPlanError::FusedGroupTransportOutside {
-                    group_id: group.group_id,
-                    transport_id,
-                });
-            }
-        }
-    }
-
-    // ── Epoch coverage: every loop in exactly one epoch, epoch loops known.
-    let mut epoch_of_loop: AHashMap<u64, u64> = AHashMap::new();
-    for epoch in &plan.epochs {
-        strictly_ascending(&epoch.loops).map_err(|at| VectorPlanError::NotCanonical {
-            what: "epoch.loops",
-            at,
-        })?;
-        for &l in &epoch.loops {
-            if !loop_set.contains(&l) {
-                return Err(VectorPlanError::EpochLoopUnknown {
-                    epoch_id: epoch.epoch_id,
-                    loop_id: l,
-                });
-            }
-            if epoch_of_loop.insert(l, epoch.epoch_id).is_some() {
-                return Err(VectorPlanError::EpochCoverageMismatch { loop_id: l });
-            }
-        }
-    }
-
-    for &l in &loop_ids {
-        if !epoch_of_loop.contains_key(&l) {
-            return Err(VectorPlanError::EpochCoverageMismatch { loop_id: l });
-        }
-    }
-    for lp in &plan.loops {
-        let actual = epoch_of_loop[&lp.loop_id];
-        if lp.epoch_id != actual {
-            return Err(VectorPlanError::LoopEpochMismatch {
-                loop_id: lp.loop_id,
-                declared: lp.epoch_id,
-                actual,
-            });
-        }
-    }
-
-    // ── Placement / roots agreement (P-Unique, P-Root, P-Duplicate). ─────
-    for sig in &plan.signals {
-        let derived_duplicable = effects_duplicable(&sig.effects);
-        if sig.duplicable != derived_duplicable {
-            return Err(VectorPlanError::DuplicabilityMismatch {
-                signal_id: sig.signal_id,
-            });
-        }
-        if sig.structural
-            && (sig.placement != Placement::Inline
-                || !matches!(sig.value_type, ValueType::Tuple(_)))
-        {
-            return Err(VectorPlanError::InlineNotDuplicable {
-                signal_id: sig.signal_id,
-            });
-        }
-        if sig.placement == Placement::Inline && !derived_duplicable && !sig.structural {
-            return Err(VectorPlanError::InlineNotDuplicable {
-                signal_id: sig.signal_id,
-            });
-        }
-        if let Placement::Owned(l) = sig.placement {
-            let owner = loop_by_id
-                .get(&l)
-                .ok_or(VectorPlanError::OwnedSignalNotRoot {
-                    signal_id: sig.signal_id,
-                    loop_id: l,
-                })?;
-            if !owner.roots.contains(&sig.signal_id) {
-                return Err(VectorPlanError::OwnedSignalNotRoot {
-                    signal_id: sig.signal_id,
-                    loop_id: l,
-                });
-            }
-        }
-    }
-    for lp in &plan.loops {
-        // Roots must be unique within a loop (Lean `rootsNodup`) but need not
-        // be ascending (deterministic materialization order, not a set-like
-        // canonical array). Uniqueness and ownership are checked together.
-        let mut seen = AHashSet::new();
-        for &r in &lp.roots {
-            if !signal_set.contains(&r) {
-                return Err(VectorPlanError::RootUnknownSignal {
-                    loop_id: lp.loop_id,
-                    signal_id: r,
-                });
-            }
-            if !seen.insert(r) {
-                return Err(VectorPlanError::RootWithoutOwnership {
-                    signal_id: r,
-                    loop_id: lp.loop_id,
-                });
-            }
-            match signal_by_id.get(&r).map(|s| s.placement) {
-                Some(Placement::Owned(owner)) if owner == lp.loop_id => {}
-                _ => {
-                    return Err(VectorPlanError::RootWithoutOwnership {
-                        signal_id: r,
-                        loop_id: lp.loop_id,
-                    });
-                }
-            }
-        }
-    }
-
-    // ── Edges: endpoints exist, no self-edge, barriers monotone. ─────────
-    for edge in plan.data_edges.iter().chain(plan.effect_edges.iter()) {
-        if !loop_set.contains(&edge.consumer) {
-            return Err(VectorPlanError::EdgeEndpointUnknown {
-                edge: *edge,
-                missing: edge.consumer,
-            });
-        }
-        if !loop_set.contains(&edge.dependency) {
-            return Err(VectorPlanError::EdgeEndpointUnknown {
-                edge: *edge,
-                missing: edge.dependency,
-            });
-        }
-        if edge.consumer == edge.dependency {
-            return Err(VectorPlanError::LoopSelfEdge {
-                loop_id: edge.consumer,
-            });
-        }
-        // Barrier: dependency epoch rank ≤ consumer epoch rank.
-        let dep_rank = rank_of(plan, epoch_of_loop[&edge.dependency]);
-        let con_rank = rank_of(plan, epoch_of_loop[&edge.consumer]);
-        if dep_rank > con_rank {
-            return Err(VectorPlanError::BarrierViolation { edge: *edge });
-        }
-    }
-
-    // ── Per-epoch induced-graph acyclicity (L-DAG). ──────────────────────
-    for epoch in &plan.epochs {
-        let members: AHashSet<u64> = epoch.loops.iter().copied().collect();
-        if let Some(remaining) = induced_cycle(plan, &members) {
-            return Err(VectorPlanError::EpochNotAcyclic {
-                epoch_id: epoch.epoch_id,
-                remaining,
-            });
-        }
-    }
-
-    // ── Effect conflicts: every conflicting loop pair is comparable. ──
-    // Root identities and graph endpoints are known valid at this point, so
-    // this check cannot panic when presented with a hostile DTO.
-    let effects_by_loop = plan
-        .loops
-        .iter()
-        .map(|loop_record| {
-            (
-                loop_record.loop_id,
-                CheckedEffectConflictSummary::new(&signal_by_id, loop_record),
-            )
-        })
-        .collect::<AHashMap<_, _>>();
-    let reachability = CheckedReachability::new(plan);
-    for (index, left) in plan.loops.iter().enumerate() {
-        for right in &plan.loops[index + 1..] {
-            if effects_by_loop[&left.loop_id].conflicts(&effects_by_loop[&right.loop_id])
-                && !reachability.reaches(left.loop_id, right.loop_id)
-                && !reachability.reaches(right.loop_id, left.loop_id)
-            {
-                return Err(VectorPlanError::UnorderedEffectConflict {
-                    left: left.loop_id,
-                    right: right.loop_id,
-                });
-            }
-        }
-    }
-
-    // ── Transports well-typed (T-TRANSPORT). ─────────────────────────────
-    for t in &plan.transports {
-        if !signal_set.contains(&t.signal_id) {
-            return Err(VectorPlanError::TransportUnknownRef {
-                transport_id: t.transport_id,
-                missing: t.signal_id,
-            });
-        }
-        for l in [t.producer_loop, t.consumer_loop] {
-            if !loop_set.contains(&l) {
-                return Err(VectorPlanError::TransportUnknownRef {
-                    transport_id: t.transport_id,
-                    missing: l,
-                });
-            }
-        }
-        if t.producer_loop == t.consumer_loop {
-            return Err(VectorPlanError::TransportSelfLoop {
-                transport_id: t.transport_id,
-            });
-        }
-        if signal_by_id[&t.signal_id].value_type != t.element_type {
-            return Err(VectorPlanError::TransportTypeMismatch {
-                transport_id: t.transport_id,
-            });
-        }
-        if t.length != plan.vec_size {
-            return Err(VectorPlanError::TransportLengthMismatch {
-                transport_id: t.transport_id,
-            });
-        }
-        if let TransportLayout::Interleaved(width) = t.layout {
-            let matching_bundle = plan.lockstep_bundles.iter().any(|bundle| {
-                u64::try_from(bundle.member_loop_ids.len()).ok() == Some(width)
-                    && (bundle
-                        .member_loop_ids
-                        .binary_search(&t.producer_loop)
-                        .is_ok()
-                        || bundle
-                            .member_loop_ids
-                            .binary_search(&t.consumer_loop)
-                            .is_ok())
-            });
-            if width < 2 || !matching_bundle {
-                return Err(VectorPlanError::TransportLayoutMismatch {
-                    transport_id: t.transport_id,
-                });
-            }
-        }
-    }
-
-    // ── VecSafe witnesses vs loop kinds. ─────────────────────────────────
-    let mut witness_of: AHashMap<u64, WitnessKind> = AHashMap::new();
-    for w in &plan.vec_safe_witnesses {
-        if !loop_set.contains(&w.loop_id) {
-            return Err(VectorPlanError::WitnessUnknownLoop { loop_id: w.loop_id });
-        }
-        witness_of.insert(w.loop_id, w.witness_kind);
-    }
-    for lp in &plan.loops {
-        match lp.kind {
-            LoopKind::Vectorizable => {
-                let vec_safe = lp.roots.iter().all(|root| {
-                    let signal = signal_by_id[root];
-                    signal.vectorability == Vectorability::Vect
-                        && effects_sample_reorderable(&signal.effects)
-                });
-                if !vec_safe {
-                    return Err(VectorPlanError::VectorizableNotSafe {
-                        loop_id: lp.loop_id,
-                    });
-                }
-                let Some(kind) = witness_of.get(&lp.loop_id) else {
-                    return Err(VectorPlanError::VectorizableWithoutWitness {
-                        loop_id: lp.loop_id,
-                    });
-                };
-                // A vectorizable loop's witness must be a *vectorizing* one.
-                if !matches!(kind, WitnessKind::Pointwise | WitnessKind::ProvenIntrinsic) {
-                    return Err(VectorPlanError::VectorizableWithoutWitness {
-                        loop_id: lp.loop_id,
-                    });
-                }
-            }
-            LoopKind::Recursive(_) | LoopKind::Island(_) | LoopKind::Lockstep { .. } => {
-                // Serial loops must not claim a pointwise (per-lane
-                // parallel) witness; only a serial-state witness is
-                // consistent with their kind.
-                if matches!(witness_of.get(&lp.loop_id), Some(WitnessKind::Pointwise)) {
-                    return Err(VectorPlanError::SerialLoopNotSerial {
-                        loop_id: lp.loop_id,
-                    });
-                }
-            }
-        }
-    }
-
-    Ok(())
-}
-
-/// Independently verifies the decoration-backed obligations of every fused
-/// serial group.
-///
-/// JSON Schema and [`verify_vector_plan`] can validate only the finite shape
-/// and plan-local identities. This second L2 gate reconstructs recursion,
-/// delay, and clock facts from an already accepted decoration certificate; it
-/// never calls the vector-plan producer.
-pub fn verify_fused_serial_groups(
-    plan: &VectorPlan,
-    decorations: &VerifiedDecorationCertificate,
-) -> Result<(), VectorPlanError> {
-    verify_vector_plan(plan)?;
-    verify_fused_serial_groups_after_plan(plan, decorations)
-}
-
-/// Verifies fused-group obligations after the caller has already accepted the
-/// same plan with [`verify_vector_plan`]. This avoids repeating the expensive
-/// independent plan check at the production boundary while preserving the
-/// standalone public checker's fail-closed contract.
-pub(crate) fn verify_fused_serial_groups_after_plan(
-    plan: &VectorPlan,
-    decorations: &VerifiedDecorationCertificate,
-) -> Result<(), VectorPlanError> {
-    let certificate = decorations.certificate();
-    let records = certificate
-        .records
-        .iter()
-        .map(|record| (u64::from(record.signal_id), record))
-        .collect::<AHashMap<_, _>>();
-    let signals = plan
-        .signals
-        .iter()
-        .map(|signal| (signal.signal_id, signal))
-        .collect::<AHashMap<_, _>>();
-    let loops = plan
-        .loops
-        .iter()
-        .map(|loop_| (loop_.loop_id, loop_))
-        .collect::<AHashMap<_, _>>();
-    let reachability = CheckedReachability::new(plan);
-    let delayed_occurrences = certificate
-        .occurrence_dependencies
-        .iter()
-        .filter_map(|dependency| {
-            (dependency.delay > 0).then_some((u64::from(dependency.from), u64::from(dependency.to)))
-        })
-        .collect::<AHashSet<_>>();
-    for group in &plan.fused_serial_groups {
-        let mut carrier_owners = BTreeSet::new();
-        for &carrier_id in &group.state_carrier_signal_ids {
-            let Some(carrier) = records.get(&carrier_id).copied() else {
-                return Err(VectorPlanError::FusedGroupCarrierNotDelayedState {
-                    group_id: group.group_id,
-                    signal_id: carrier_id,
-                });
-            };
-            let Placement::Owned(owner_loop_id) = signals[&carrier_id].placement else {
-                return Err(VectorPlanError::FusedGroupCarrierNotDelayedState {
-                    group_id: group.group_id,
-                    signal_id: carrier_id,
-                });
-            };
-            if carrier.max_delay == 0
-                || group.member_loop_ids.binary_search(&owner_loop_id).is_err()
-                || group
-                    .state_write_signal_ids
-                    .binary_search(&carrier_id)
-                    .is_err()
-            {
-                return Err(VectorPlanError::FusedGroupCarrierNotDelayedState {
-                    group_id: group.group_id,
-                    signal_id: carrier_id,
-                });
-            }
-            if let Some(projection) = carrier.recursive_projection
-                && loops[&owner_loop_id].kind != LoopKind::Recursive(u64::from(projection.group))
-            {
-                return Err(VectorPlanError::FusedGroupCarrierNotDelayedState {
-                    group_id: group.group_id,
-                    signal_id: carrier_id,
-                });
-            }
-            carrier_owners.insert(owner_loop_id);
-        }
-        if carrier_owners.first().copied() != Some(group.owner_loop_id) {
-            return Err(VectorPlanError::FusedGroupOwnerNotStateCarrier {
-                group_id: group.group_id,
-                owner_loop_id: group.owner_loop_id,
-            });
-        }
-        for (&signal_id, record) in &records {
-            let Some(signal) = signals.get(&signal_id) else {
-                continue;
-            };
-            let Placement::Owned(owner_loop_id) = signal.placement else {
-                continue;
-            };
-            if record.max_delay == 0 || group.member_loop_ids.binary_search(&owner_loop_id).is_err()
-            {
-                continue;
-            }
-            let has_owned_group_read =
-                certificate.dependencies.iter().any(|dependency| {
-                    u64::from(dependency.to) == signal_id
-                    && (matches!(dependency.kind, DepKind::Delayed { amount } if amount > 0)
-                        || (matches!(dependency.kind, DepKind::Immediate)
-                            && delayed_occurrences
-                                .contains(&(u64::from(dependency.from), signal_id))))
-                    && signals
-                        .get(&u64::from(dependency.from))
-                        .is_some_and(|read| matches!(
-                            read.placement,
-                            Placement::Owned(read_loop_id)
-                                if group.member_loop_ids.binary_search(&read_loop_id).is_ok()
-                        ))
-                });
-            if has_owned_group_read
-                && group
-                    .state_carrier_signal_ids
-                    .binary_search(&signal_id)
-                    .is_err()
-            {
-                return Err(VectorPlanError::FusedGroupCarrierNotDelayedState {
-                    group_id: group.group_id,
-                    signal_id,
-                });
-            }
-        }
-
-        let mut group_clock = None;
-        for signal in plan.signals.iter().filter(|signal| {
-            matches!(signal.placement, Placement::Owned(owner) if group.member_loop_ids.binary_search(&owner).is_ok())
-        }) {
-            if group_clock.replace(signal.clock_id).is_some_and(|clock| clock != signal.clock_id) {
-                return Err(VectorPlanError::FusedGroupClockMismatch {
-                    group_id: group.group_id,
-                });
-            }
-        }
-        for &signal_id in group
-            .state_carrier_signal_ids
-            .iter()
-            .chain(&group.delayed_read_signal_ids)
-            .chain(&group.state_write_signal_ids)
-            .chain(&group.output_or_transport_roots)
-        {
-            let Some(record) = records.get(&signal_id).copied() else {
-                return Err(VectorPlanError::FusedGroupUnknownSignal {
-                    group_id: group.group_id,
-                    signal_id,
-                });
-            };
-            let decoration_clock = record.clock_domain.map_or(0, |clock| u64::from(clock) + 1);
-            if decoration_clock != signals[&signal_id].clock_id
-                || group_clock.is_some_and(|clock| clock != decoration_clock)
-            {
-                return Err(VectorPlanError::FusedGroupClockMismatch {
-                    group_id: group.group_id,
-                });
-            }
-        }
-
-        let mut used_carriers = BTreeSet::new();
-        let mut immediate_state_effects = BTreeMap::<u64, BTreeSet<EffectAtom>>::new();
-        for &read_signal_id in &group.delayed_read_signal_ids {
-            let delayed_carrier_edges = certificate
-                .dependencies
-                .iter()
-                .filter(|dependency| {
-                    u64::from(dependency.from) == read_signal_id
-                        && group
-                            .state_carrier_signal_ids
-                            .binary_search(&u64::from(dependency.to))
-                            .is_ok()
-                        && (matches!(dependency.kind, DepKind::Delayed { amount } if amount > 0)
-                            || (matches!(dependency.kind, DepKind::Immediate)
-                                && delayed_occurrences
-                                    .contains(&(read_signal_id, u64::from(dependency.to)))))
-                })
-                .collect::<Vec<_>>();
-            if delayed_carrier_edges.is_empty() {
-                if std::env::var_os("FAUST_RS_VECTOR_TIMING").is_some() {
-                    eprintln!(
-                        "[vector-fused-unmatched-read] group={} read={} carriers={:?} dependencies={:?}",
-                        group.group_id,
-                        read_signal_id,
-                        group.state_carrier_signal_ids,
-                        certificate
-                            .dependencies
-                            .iter()
-                            .filter(|dependency| { u64::from(dependency.from) == read_signal_id })
-                            .collect::<Vec<_>>()
-                    );
-                }
-                return Err(VectorPlanError::FusedGroupDelayedDependencyMissing {
-                    group_id: group.group_id,
-                    signal_id: read_signal_id,
-                });
-            }
-            let Placement::Owned(read_loop_id) = signals[&read_signal_id].placement else {
-                return Err(VectorPlanError::FusedGroupSignalOutside {
-                    group_id: group.group_id,
-                    signal_id: read_signal_id,
-                });
-            };
-            for delayed_carrier_edge in delayed_carrier_edges {
-                let carrier_signal_id = u64::from(delayed_carrier_edge.to);
-                used_carriers.insert(carrier_signal_id);
-                if matches!(delayed_carrier_edge.kind, DepKind::Immediate) {
-                    immediate_state_effects
-                        .entry(carrier_signal_id)
-                        .or_default()
-                        .extend(
-                            signals[&carrier_signal_id]
-                                .effects
-                                .iter()
-                                .filter(|carrier_effect| {
-                                    signals[&read_signal_id].effects.iter().any(|read_effect| {
-                                        effects_conflict(carrier_effect, read_effect)
-                                    })
-                                })
-                                .cloned(),
-                        );
-                }
-                let Placement::Owned(writer_loop_id) = signals[&carrier_signal_id].placement else {
-                    return Err(VectorPlanError::FusedGroupSignalOutside {
-                        group_id: group.group_id,
-                        signal_id: carrier_signal_id,
-                    });
-                };
-                let (path_start, path_end) =
-                    if matches!(delayed_carrier_edge.kind, DepKind::Delayed { .. }) {
-                        (read_loop_id, writer_loop_id)
-                    } else {
-                        (writer_loop_id, read_loop_id)
-                    };
-                for &loop_id in loops.keys() {
-                    let follows_start =
-                        loop_id == path_start || reachability.reaches(path_start, loop_id);
-                    let precedes_end =
-                        loop_id == path_end || reachability.reaches(loop_id, path_end);
-                    if follows_start
-                        && precedes_end
-                        && group.member_loop_ids.binary_search(&loop_id).is_err()
-                    {
-                        return Err(VectorPlanError::FusedGroupPathIncomplete {
-                            group_id: group.group_id,
-                            loop_id,
-                        });
-                    }
-                }
-            }
-        }
-        if used_carriers
-            != group
-                .state_carrier_signal_ids
-                .iter()
-                .copied()
-                .collect::<BTreeSet<_>>()
-        {
-            let signal_id = group
-                .state_carrier_signal_ids
-                .iter()
-                .find(|signal_id| !used_carriers.contains(signal_id))
-                .copied()
-                .unwrap_or(u64::MAX);
-            return Err(VectorPlanError::FusedGroupCarrierNotDelayedState {
-                group_id: group.group_id,
-                signal_id,
-            });
-        }
-        for state_effects in immediate_state_effects.values() {
-            for loop_record in plan.loops.iter().filter(|loop_record| {
-                loop_record.roots.iter().any(|root| {
-                    signals[root].effects.iter().any(|effect| {
-                        state_effects
-                            .iter()
-                            .any(|carrier| effects_conflict(carrier, effect))
-                    })
-                })
-            }) {
-                if group
-                    .member_loop_ids
-                    .binary_search(&loop_record.loop_id)
-                    .is_err()
-                {
-                    return Err(VectorPlanError::FusedGroupPathIncomplete {
-                        group_id: group.group_id,
-                        loop_id: loop_record.loop_id,
-                    });
-                }
-            }
-        }
-        let group_effects = group
-            .member_loop_ids
-            .iter()
-            .flat_map(|loop_id| loops[loop_id].roots.iter())
-            .flat_map(|root| signals[root].effects.iter())
-            .cloned()
-            .collect::<BTreeSet<_>>();
-        for loop_record in plan.loops.iter().filter(|loop_record| {
-            group
-                .member_loop_ids
-                .binary_search(&loop_record.loop_id)
-                .is_err()
-        }) {
-            let conflicts = loop_record.roots.iter().any(|root| {
-                signals[root].effects.iter().any(|effect| {
-                    group_effects
-                        .iter()
-                        .any(|group_effect| effects_conflict(group_effect, effect))
-                })
-            });
-            if conflicts {
-                return Err(VectorPlanError::FusedGroupPathIncomplete {
-                    group_id: group.group_id,
-                    loop_id: loop_record.loop_id,
-                });
-            }
-        }
-        for &loop_id in loops.keys() {
-            if group.member_loop_ids.binary_search(&loop_id).is_ok() {
-                continue;
-            }
-            let follows_member = group
-                .member_loop_ids
-                .iter()
-                .any(|member| reachability.reaches(*member, loop_id));
-            let precedes_member = group
-                .member_loop_ids
-                .iter()
-                .any(|member| reachability.reaches(loop_id, *member));
-            if follows_member && precedes_member {
-                return Err(VectorPlanError::FusedGroupPathIncomplete {
-                    group_id: group.group_id,
-                    loop_id,
-                });
-            }
-        }
-
-        for &writer_signal_id in &group.state_write_signal_ids {
-            if group
-                .state_carrier_signal_ids
-                .binary_search(&writer_signal_id)
-                .is_ok()
-            {
-                continue;
-            }
-            let Some(writer_projection) = records
-                .get(&writer_signal_id)
-                .and_then(|record| record.recursive_projection)
-            else {
-                return Err(VectorPlanError::FusedGroupStateWriterMismatch {
-                    group_id: group.group_id,
-                    signal_id: writer_signal_id,
-                });
-            };
-            let Placement::Owned(writer_owner) = signals[&writer_signal_id].placement else {
-                return Err(VectorPlanError::FusedGroupStateWriterMismatch {
-                    group_id: group.group_id,
-                    signal_id: writer_signal_id,
-                });
-            };
-            if group.member_loop_ids.binary_search(&writer_owner).is_err()
-                || loops[&writer_owner].kind
-                    != LoopKind::Recursive(u64::from(writer_projection.group))
-            {
-                return Err(VectorPlanError::FusedGroupStateWriterMismatch {
-                    group_id: group.group_id,
-                    signal_id: writer_signal_id,
-                });
-            }
-        }
-        for &loop_id in &group.member_loop_ids {
-            let LoopKind::Recursive(recursion_group) = loops[&loop_id].kind else {
-                continue;
-            };
-            if !group.state_write_signal_ids.iter().any(|signal_id| {
-                signals[signal_id].placement == Placement::Owned(loop_id)
-                    && records[signal_id]
-                        .recursive_projection
-                        .is_some_and(|projection| u64::from(projection.group) == recursion_group)
-            }) {
-                if std::env::var_os("FAUST_RS_VECTOR_TIMING").is_some() {
-                    eprintln!(
-                        "[vector-fused-missing-writer] group={} loop={} recursion={} writers={:?}",
-                        group.group_id, loop_id, recursion_group, group.state_write_signal_ids
-                    );
-                }
-                return Err(VectorPlanError::FusedGroupRecursiveMemberMissingWriter {
-                    group_id: group.group_id,
-                    loop_id,
-                });
-            }
-        }
-
-        // Internal transports may form an arbitrary pure chain between a
-        // delayed read and its recursive writer. Finite-shape verification
-        // above proves that every listed transport stays within the fused
-        // member set; require every internal transport so none can remain a
-        // whole-chunk array inside the serial slice.
-        for transport in &plan.transports {
-            if group
-                .member_loop_ids
-                .binary_search(&transport.producer_loop)
-                .is_ok()
-                && group
-                    .member_loop_ids
-                    .binary_search(&transport.consumer_loop)
-                    .is_ok()
-                && group
-                    .internal_transport_ids
-                    .binary_search(&transport.transport_id)
-                    .is_err()
-            {
-                return Err(VectorPlanError::FusedGroupDangerousTransportPresent {
-                    group_id: group.group_id,
-                    transport_id: transport.transport_id,
-                });
-            }
-        }
-    }
-
-    // Reconstruct the owned subset of immediate state-mediated crossings from
-    // decorations and raw plan placement. This is deliberately independent of
-    // producer component discovery and prevents removing the producer's final
-    // fail-closed edge guard without equivalent certificate coverage.
-    for dependency in certificate.dependencies.iter().filter(|dependency| {
-        matches!(dependency.kind, DepKind::Immediate)
-            && delayed_occurrences.contains(&(u64::from(dependency.from), u64::from(dependency.to)))
-            && records
-                .get(&u64::from(dependency.from))
-                .is_some_and(|record| record.is_delay_read)
-            && records
-                .get(&u64::from(dependency.to))
-                .is_some_and(|record| record.max_delay > 0)
-    }) {
-        let read_signal_id = u64::from(dependency.from);
-        let carrier_signal_id = u64::from(dependency.to);
-        let (Placement::Owned(consumer), Placement::Owned(producer)) = (
-            signals[&read_signal_id].placement,
-            signals[&carrier_signal_id].placement,
-        ) else {
-            continue;
-        };
-        if producer == consumer {
-            continue;
-        }
-        let covered = plan.fused_serial_groups.iter().any(|group| {
-            group
-                .state_carrier_signal_ids
-                .binary_search(&carrier_signal_id)
-                .is_ok()
-                && group
-                    .delayed_read_signal_ids
-                    .binary_search(&read_signal_id)
-                    .is_ok()
-                && group.member_loop_ids.binary_search(&producer).is_ok()
-                && group.member_loop_ids.binary_search(&consumer).is_ok()
-        });
-        if !covered {
-            return Err(VectorPlanError::FusedGroupDangerousCrossingMissing { producer, consumer });
-        }
-    }
-    Ok(())
-}
-
-fn rank_of(plan: &VectorPlan, epoch_id: u64) -> u64 {
-    plan.epochs
-        .iter()
-        .find(|e| e.epoch_id == epoch_id)
-        .map_or(u64::MAX, |e| e.rank)
-}
-
-struct CheckedEffectConflictSummary {
-    any: bool,
-    barrier: bool,
-    state_reads: AHashSet<StateResource>,
-    state_writes: AHashSet<StateResource>,
-    table_reads: AHashSet<u32>,
-    table_writes: AHashSet<u32>,
-    ui_writes: AHashSet<u32>,
-    output_writes: AHashSet<u32>,
-}
-
-impl Default for CheckedEffectConflictSummary {
-    fn default() -> Self {
-        Self {
-            any: false,
-            barrier: false,
-            state_reads: AHashSet::new(),
-            state_writes: AHashSet::new(),
-            table_reads: AHashSet::new(),
-            table_writes: AHashSet::new(),
-            ui_writes: AHashSet::new(),
-            output_writes: AHashSet::new(),
-        }
-    }
-}
-
-impl CheckedEffectConflictSummary {
-    fn new(signal_by_id: &AHashMap<u64, &SignalRecord>, loop_record: &LoopRecord) -> Self {
-        let mut summary = Self::default();
-        for effect in loop_record
-            .roots
-            .iter()
-            .flat_map(|root| &signal_by_id[root].effects)
-        {
-            summary.any = true;
-            match effect {
-                EffectAtom::ReadState(resource) => {
-                    summary.state_reads.insert(resource.clone());
-                }
-                EffectAtom::WriteState(resource) => {
-                    summary.state_writes.insert(resource.clone());
-                }
-                EffectAtom::ReadTable(table) => {
-                    summary.table_reads.insert(*table);
-                }
-                EffectAtom::WriteTable(table) => {
-                    summary.table_writes.insert(*table);
-                }
-                EffectAtom::WriteUi(zone) => {
-                    summary.ui_writes.insert(*zone);
-                }
-                EffectAtom::WriteOutput(output) => {
-                    summary.output_writes.insert(*output);
-                }
-                EffectAtom::Foreign { purity, .. } => {
-                    summary.barrier |=
-                        matches!(purity, ForeignPurity::Impure | ForeignPurity::Unknown);
-                }
-            }
-        }
-        summary
-    }
-
-    fn conflicts(&self, other: &Self) -> bool {
-        (self.barrier && other.any)
-            || (other.barrier && self.any)
-            || hash_intersects(&self.state_writes, &other.state_reads)
-            || hash_intersects(&self.state_writes, &other.state_writes)
-            || hash_intersects(&self.state_reads, &other.state_writes)
-            || hash_intersects(&self.table_writes, &other.table_reads)
-            || hash_intersects(&self.table_writes, &other.table_writes)
-            || hash_intersects(&self.table_reads, &other.table_writes)
-            || hash_intersects(&self.ui_writes, &other.ui_writes)
-            || hash_intersects(&self.output_writes, &other.output_writes)
-    }
-}
-
-fn hash_intersects<T: Eq + std::hash::Hash>(left: &AHashSet<T>, right: &AHashSet<T>) -> bool {
-    let (small, large) = if left.len() <= right.len() {
-        (left, right)
-    } else {
-        (right, left)
-    };
-    small.iter().any(|item| large.contains(item))
-}
-
-/// Checker-local transitive closure. This deliberately does not reuse the
-/// producer's implementation: the certificate boundary remains independent,
-/// while avoiding a BFS and signal-map rebuild for every loop pair.
-struct CheckedReachability {
-    index: AHashMap<u64, usize>,
-    rows: Vec<Vec<u64>>,
-}
-
-impl CheckedReachability {
-    fn new(plan: &VectorPlan) -> Self {
-        Self::from_edges(plan, plan.data_edges.iter().chain(&plan.effect_edges))
-    }
-
-    fn from_edges<'a>(plan: &VectorPlan, edges: impl Iterator<Item = &'a LoopEdge>) -> Self {
-        let index = plan
-            .loops
-            .iter()
-            .enumerate()
-            .map(|(index, loop_record)| (loop_record.loop_id, index))
-            .collect::<AHashMap<_, _>>();
-        let words = plan.loops.len().div_ceil(u64::BITS as usize);
-        let mut rows = vec![vec![0_u64; words]; plan.loops.len()];
-        for edge in edges {
-            let from = index[&edge.dependency];
-            let to = index[&edge.consumer];
-            rows[from][to / u64::BITS as usize] |= 1_u64 << (to % u64::BITS as usize);
-        }
-        for intermediate in 0..rows.len() {
-            let additions = rows[intermediate].clone();
-            for row in &mut rows {
-                if row[intermediate / u64::BITS as usize]
-                    & (1_u64 << (intermediate % u64::BITS as usize))
-                    != 0
-                {
-                    for (target, addition) in row.iter_mut().zip(&additions) {
-                        *target |= addition;
-                    }
-                }
-            }
-        }
-        Self { index, rows }
-    }
-
-    fn reaches(&self, from: u64, to: u64) -> bool {
-        let from = self.index[&from];
-        let to = self.index[&to];
-        self.rows[from][to / u64::BITS as usize] & (1_u64 << (to % u64::BITS as usize)) != 0
-    }
-}
-
-/// Kahn peeling on the induced graph of `members` (edges from both edge
-/// families whose endpoints are both members). Returns the unschedulable set
-/// (stable-sorted) if a cycle remains, else `None`.
-fn induced_cycle(plan: &VectorPlan, members: &AHashSet<u64>) -> Option<Vec<u64>> {
-    let mut pending: AHashMap<u64, AHashSet<u64>> = AHashMap::new();
-    let mut successors: AHashMap<u64, Vec<u64>> = AHashMap::new();
-    for &m in members {
-        pending.insert(m, AHashSet::new());
-    }
-    for edge in plan.data_edges.iter().chain(plan.effect_edges.iter()) {
-        if members.contains(&edge.consumer) && members.contains(&edge.dependency) {
-            pending
-                .get_mut(&edge.consumer)
-                .expect("consumer is a member")
-                .insert(edge.dependency);
-            successors
-                .entry(edge.dependency)
-                .or_default()
-                .push(edge.consumer);
-        }
-    }
-    let mut ready: Vec<u64> = pending
-        .iter()
-        .filter(|(_, deps)| deps.is_empty())
-        .map(|(&n, _)| n)
-        .collect();
-    let mut removed: AHashSet<u64> = AHashSet::new();
-    while let Some(n) = ready.pop() {
-        if !removed.insert(n) {
-            continue;
-        }
-        if let Some(succs) = successors.get(&n) {
-            for &s in succs {
-                if let Some(set) = pending.get_mut(&s) {
-                    set.remove(&n);
-                    if set.is_empty() && !removed.contains(&s) {
-                        ready.push(s);
-                    }
-                }
-            }
-        }
-    }
-    if removed.len() == members.len() {
-        None
-    } else {
-        let mut remaining: Vec<u64> = members.difference(&removed).copied().collect();
-        remaining.sort_unstable();
-        Some(remaining)
-    }
-}
+pub use check::*;
+pub(crate) use check::{effects_duplicable, effects_sample_reorderable};
+pub use error::*;
+pub use fused_groups::*;
+pub use model::*;
 
 #[cfg(test)]
 mod tests;
