+++ porting/journal/2026-07-19.md
@@ -1,5 +1,29 @@
 # 2026-07-19
 
+## Split the vector plan verifier by responsibility (R5.2)
+
+Commit date: 2026-07-19
+Source commit: this commit
+
+`vector/verify/mod.rs` (1,983 production lines) becomes five files behind an
+unchanged public facade (`pub use` re-exports keep every old path,
+including `verify::EffectAtom` and the `pub(crate)` domain axioms):
+
+- `model.rs` (206) — the canonical schema-v4 `VectorPlan` DTO vocabulary,
+  pure data, serialization order and schema version frozen;
+- `error.rs` (376) — `VectorPlanError` and its `Display`/`Error` impls;
+- `check.rs` (835) — `verify_vector_plan`, the `effects_duplicable` /
+  `effects_sample_reorderable` domain axioms, and checker-internal helpers
+  (`CheckedEffectConflictSummary`, `rank_of`, `induced_cycle`, ...);
+- `fused_groups.rs` (476) — `verify_fused_serial_groups(_after_plan)`;
+- `checker_reachability.rs` (54) — `CheckedReachability`, explicitly
+  documented as independent of the producer's reachability (trust boundary).
+
+Move only: bodies are byte-identical apart from `pub(super)` on helpers now
+crossed by sibling files and per-file imports. Every resulting production
+file is under the plan's 1,200-line threshold. Validation: transform 385,
+`vector_mode` oracle 35/35, clippy `-D warnings` clean, fmt applied.
+
 ## Seed the vector vocabulary module with prepared-ID indexing (R4.2, part 1)
 
 Commit date: 2026-07-19
